♻️ Split the proc-macro implementation into a dedicated yuuka-macros crate.
- Turn the repository into a Cargo workspace: the root stays the yuuka
  facade package which re-exports yuuka_macros, so yuuka::derive_struct!,
  yuuka::derive_enum! and yuuka::auto! paths are unchanged.
- Move the whole macro implementation tree to macros/ as the yuuka-macros
  proc-macro crate and bump the shared workspace version to 0.7.0.
- Drop unused macro-crate dependencies async-trait, derive_more, serde,
  serde_json and strum plus the unused span-locations feature of
  proc-macro2; keep only anyhow, proc-macro2, quote and syn.
- Move serde, serde_json and anyhow to facade dev-dependencies where the
  integration tests and examples actually use them.
- Remove an unused ToTokens import so clippy stays warning free.
- Add macros path filters to CI workflows and publish yuuka-macros before
  yuuka in the crates.io release workflow.
