+++ .github/workflows/clippy.yml
@@ -6,6 +6,7 @@ on:
     paths:
       - '.github/workflows/clippy.yml'
       - 'src/**/*'
+      - 'macros/**/*'
       - 'tests/**/*'
       - 'Cargo.toml'
   push:
