+++ .github/workflows/test.yml
@@ -6,6 +6,7 @@ on:
     paths:
       - '.github/workflows/test.yml'
       - 'src/**/*'
+      - 'macros/**/*'
       - 'tests/**/*'
       - 'Cargo.toml'
   push:
