+++ macros/src/template/auto_enums.rs
