+++ macros/src/template/auto_structs.rs
