+++ macros/src/template/enums.rs
