+++ macros/src/template/structs.rs
