+++ macros/src/tools/auto_macros.rs
