+++ macros/src/tools/derive_enum.rs
