+++ macros/src/tools/derive_enum_items.rs
