+++ macros/src/utils/flatten.rs
